requirement.txt
asgiref==3.7.2
Django==4.2.2
django-autoslug==1.9.9
sqlparse==0.4.4
tzdata==2023.3
